@@ -31,31 +31,36 @@
   },
   "homepage": "https://github.com/billmalarky/react-native-image-cache-hoc#readme",
   "devDependencies": {
-    "babel-eslint": "^8.0.1",
+    "babel-eslint": "^10.1.0",
     "coveralls": "^3.0.0",
-    "eslint": "^4.9.0",
-    "eslint-plugin-react": "^7.4.0",
-    "eslint-plugin-react-native": "^3.1.0",
-    "jest": "^21.2.1",
+    "eslint": "^7.8.1",
+    "eslint-plugin-react": "^7.20.6",
+    "eslint-plugin-react-native": "^3.9.1",
+    "jest": "^26.4.2",
     "react": "^16.0.0",
-    "react-native": "^0.50.1",
-    "react-test-renderer": "^16.0.0",
+    "react-native": "^0.63.2",
+    "react-test-renderer": "^16.13.1",
     "should": "^13.1.2",
     "should-sinon": "^0.0.6",
     "sinon": "^4.1.3"
   },
   "jest": {
     "preset": "react-native",
+    "transform": {
+      "^.+\\.js$": "<rootDir>/node_modules/react-native/jest/preprocessor.js"
+    },
     "setupFiles": [
       "./tests/config.js"
     ]
+  },
+  "peerDependencies": {
+    "react-native-fs": "^2.14.1"
   },
   "dependencies": {
     "buffer": "^5.0.8",
     "crypto-js": "^3.1.9-1",
     "path": "^0.12.7",
     "prop-types": "^15.6.0",
-    "rn-fetch-blob": "0.10.13",
     "react-native-uuid": "^1.4.9",
     "traverse": "^0.6.6",
     "url-parse": "^1.2.0",
