@@ -23,16 +23,29 @@ import validator from 'validator';
 import uuid from 'react-native-uuid';
 
 export default function imageCacheHoc(Image, options = {}) {
-
   // Validate options
-  if (options.validProtocols && !Array.isArray(options.validProtocols)) { throw new Error('validProtocols option must be an array of protocol strings.'); }
-  if (options.fileHostWhitelist && !Array.isArray(options.fileHostWhitelist)) { throw new Error('fileHostWhitelist option must be an array of host strings.'); }
-  if (options.cachePruneTriggerLimit && !Number.isInteger(options.cachePruneTriggerLimit) ) { throw new Error('cachePruneTriggerLimit option must be an integer.'); }
-  if (options.fileDirName && typeof options.fileDirName !== 'string') { throw new Error('fileDirName option must be string'); }
-  if (options.defaultPlaceholder && (!options.defaultPlaceholder.component || !options.defaultPlaceholder.props)) { throw new Error('defaultPlaceholder option object must include "component" and "props" properties (props can be an empty object)'); }
+  if (options.validProtocols && !Array.isArray(options.validProtocols)) {
+    throw new Error('validProtocols option must be an array of protocol strings.');
+  }
+  if (options.fileHostWhitelist && !Array.isArray(options.fileHostWhitelist)) {
+    throw new Error('fileHostWhitelist option must be an array of host strings.');
+  }
+  if (options.cachePruneTriggerLimit && !Number.isInteger(options.cachePruneTriggerLimit)) {
+    throw new Error('cachePruneTriggerLimit option must be an integer.');
+  }
+  if (options.fileDirName && typeof options.fileDirName !== 'string') {
+    throw new Error('fileDirName option must be string');
+  }
+  if (
+    options.defaultPlaceholder &&
+    (!options.defaultPlaceholder.component || !options.defaultPlaceholder.props)
+  ) {
+    throw new Error(
+      'defaultPlaceholder option object must include "component" and "props" properties (props can be an empty object)'
+    );
+  }
 
   return class extends React.PureComponent {
-
     static propTypes = {
       fileHostWhitelist: PropTypes.array,
       source: PropTypes.object.isRequired,
@@ -40,8 +53,9 @@ export default function imageCacheHoc(Image, options = {}) {
       style: ViewPropTypes.style,
       placeholder: PropTypes.shape({
         component: PropTypes.func,
-        props: PropTypes.object
-      })
+        props: PropTypes.object,
+      }),
+      onLoadFinished: PropTypes.func,
     };
 
     /**
@@ -56,16 +70,43 @@ export default function imageCacheHoc(Image, options = {}) {
      * @returns {Promise} promise that resolves to an object that contains cached file info.
      */
     static async cacheFile(url, permanent = false) {
-
-      const fileSystem = FileSystemFactory();
+      const fileSystem = FileSystemFactory(options.cachePruneTriggerLimit, options.fileDirName);
       const localFilePath = await fileSystem.getLocalFilePathFromUrl(url, permanent);
 
       return {
         url: url,
-        cacheType: (permanent ? 'permanent' : 'cache'),
-        localFilePath
+        cacheType: permanent ? 'permanent' : 'cache',
+        localFilePath,
       };
+    }
 
+    /**
+     *
+     * Manually cache a file.
+     * Can be used to pre-warm caches.
+     * If calling this method repeatedly to cache a long list of files,
+     * be sure to use a queue and limit concurrency so your app performance does not suffer.
+     *
+     * @param url {String} - url of local file.
+     * @param url {String} - url of file.
+     * @param permanent {Boolean} - whether the file should be saved to the tmp or permanent cache directory.
+     * @param move {Boolean} - whether the file should be moved.
+     * @returns {Promise} promise that resolves to an object that contains cached file info.
+     */
+    static async cacheLocalFile(local, url, permanent = false, move = false) {
+      const fileSystem = FileSystemFactory(options.cachePruneTriggerLimit, options.fileDirName);
+      const localFilePath = await fileSystem.getLocalFilePathFromUrlLocal(
+        local,
+        url,
+        permanent,
+        move
+      );
+
+      return {
+        url: url,
+        cacheType: permanent ? 'permanent' : 'cache',
+        localFilePath,
+      };
     }
 
     /**
@@ -76,15 +117,16 @@ export default function imageCacheHoc(Image, options = {}) {
      * @returns {Promise} promise that resolves to an object that contains the flush results.
      */
     static async flush() {
-
-      const fileSystem = FileSystemFactory();
-      const results = await Promise.all([fileSystem.unlink('permanent'), fileSystem.unlink('cache')]);
+      const fileSystem = FileSystemFactory(options.cachePruneTriggerLimit, options.fileDirName);
+      const results = await Promise.all([
+        fileSystem.unlink('permanent'),
+        fileSystem.unlink('cache'),
+      ]);
 
       return {
         permanentDirFlushed: results[0],
-        cacheDirFlushed: results[1]
+        cacheDirFlushed: results[1],
       };
-
     }
 
     constructor(props) {
@@ -92,7 +134,8 @@ export default function imageCacheHoc(Image, options = {}) {
 
       // Set initial state
       this.state = {
-        localFilePath: null
+        remoteFilePath: null,
+        localFilePath: null,
       };
 
       // Assign component unique ID for cache locking.
@@ -111,50 +154,63 @@ export default function imageCacheHoc(Image, options = {}) {
       };
 
       // Init file system lib
-      this.fileSystem = FileSystemFactory(this.options.cachePruneTriggerLimit, this.options.fileDirName);
+      this.fileSystem = FileSystemFactory(
+        this.options.cachePruneTriggerLimit,
+        this.options.fileDirName
+      );
 
       // Validate input
-      this._validateImageComponent();
-
+      this.InvalidUrl = !this._validateImageComponent();
     }
 
     _validateImageComponent() {
+      try {
+        // Define validator options
+        let validatorUrlOptions = {
+          protocols: this.options.validProtocols,
+          require_protocol: true,
+        };
+        if (this.options.fileHostWhitelist.length) {
+          validatorUrlOptions.host_whitelist = this.options.fileHostWhitelist;
+        }
 
-      // Define validator options
-      let validatorUrlOptions = { protocols: this.options.validProtocols, require_protocol: true };
-      if (this.options.fileHostWhitelist.length) {
-        validatorUrlOptions.host_whitelist = this.options.fileHostWhitelist;
+        // Validate source prop to be a valid web accessible url.
+        if (
+          !traverse(this.props).get(['source', 'uri']) ||
+          !validator.isURL(traverse(this.props).get(['source', 'uri']), validatorUrlOptions)
+        ) {
+          // throw new Error('Invalid source prop. <CacheableImage> props.source.uri should be a web accessible url with a valid protocol and host. NOTE: Default valid protocol is https, default valid hosts are *.');
+          return false;
+        } else {
+          return true;
+        }
+      } catch (e) {
+        return false;
       }
-
-      // Validate source prop to be a valid web accessible url.
-      if (
-        !traverse(this.props).get(['source', 'uri'])
-        || !validator.isURL(traverse(this.props).get(['source', 'uri']), validatorUrlOptions)
-      ) {
-        throw new Error('Invalid source prop. <CacheableImage> props.source.uri should be a web accessible url with a valid protocol and host. NOTE: Default valid protocol is https, default valid hosts are *.');
-      } else {
-        return true;
-      }
-
     }
 
     // Async calls to local FS or network should occur here.
     // See: https://reactjs.org/docs/react-component.html#componentdidmount
     async componentDidMount() {
-
       // Track component mount status to avoid calling setState() on unmounted component.
-      this._isMounted = true;
+      try {
+        this._isMounted = true;
 
-      // Set url from source prop
-      const url = traverse(this.props).get(['source', 'uri']);
+        // Set url from source prop
+        const url = traverse(this.props).get(['source', 'uri']);
 
-      // Add a cache lock to file with this name (prevents concurrent <CacheableImage> components from pruning a file with this name from cache).
-      const fileName = await this.fileSystem.getFileNameFromUrl(url);
-      FileSystem.lockCacheFile(fileName, this.componentId);
+        this.setState({ remoteFilePath: url });
 
-      // Init the image cache logic
-      await this._loadImage(url);
+        // Add a cache lock to file with this name (prevents concurrent <CacheableImage> components from pruning a file with this name from cache).
+        const fileName = await this.fileSystem.getFileNameFromUrl(url);
+        FileSystem.lockCacheFile(fileName, this.componentId);
 
+        // Init the image cache logic
+        await this._loadImage(url);
+      } catch (e) {
+        this.InvalidUrl = true;
+        this.setState({ localFilePath: null });
+      }
     }
 
     /**
@@ -162,16 +218,17 @@ export default function imageCacheHoc(Image, options = {}) {
      * Enables caching logic to work if component source prop is updated (that is, the image url changes without mounting a new component).
      * See: https://github.com/billmalarky/react-native-image-cache-hoc/pull/15
      *
-     * @param nextProps {Object} - Props that will be passed to component.
+     * @param prevProps {Object} - Previous props.
      */
-    async componentWillReceiveProps(nextProps) {
-
+    async componentDidUpdate(prevProps) {
       // Set urls from source prop data
-      const url = traverse(this.props).get(['source', 'uri']);
-      const nextUrl = traverse(nextProps).get(['source', 'uri']);
+      const url = traverse(prevProps).get(['source', 'uri']);
+      const nextUrl = traverse(this.props).get(['source', 'uri']);
 
       // Do nothing if url has not changed.
       if (url === nextUrl) return;
+
+      this.setState({ remoteFilePath: nextUrl });
 
       // Remove component cache lock on old image file, and add cache lock to new image file.
       const fileName = await this.fileSystem.getFileNameFromUrl(url);
@@ -182,7 +239,6 @@ export default function imageCacheHoc(Image, options = {}) {
 
       // Init the image cache logic
       await this._loadImage(nextUrl);
-
     }
 
     /**
@@ -194,7 +250,9 @@ export default function imageCacheHoc(Image, options = {}) {
      * @private
      */
     async _loadImage(url) {
-
+      if (this._InvalidUrl) {
+        this.setState({ localFilePath: url });
+      }
       // Check local fs for file, fallback to network and write file to disk if local file not found.
       const permanent = this.props.permanent ? true : false;
       let localFilePath = null;
@@ -204,53 +262,72 @@ export default function imageCacheHoc(Image, options = {}) {
         console.warn(error); // eslint-disable-line no-console
       }
 
+      // Check if we are still interested in the url
+      if (this.state.remoteFilePath !== url) {
+        return;
+      }
+
+      this.InvalidUrl = !localFilePath;
+
       // Check component is still mounted to avoid calling setState() on components that were quickly
       // mounted then unmounted before componentDidMount() finishes.
       // See: https://github.com/billmalarky/react-native-image-cache-hoc/issues/6#issuecomment-354490597
-      if (this._isMounted && localFilePath) {
+      if (this._isMounted) {
         this.setState({ localFilePath });
-      }
 
+        if (!this.InvalidUrl && this.props.onLoadFinished) {
+          if (Platform.OS === 'android') {
+            await Image.getSize('file://' + localFilePath, (width, height) =>
+              this.props.onLoadFinished(width, height)
+            );
+          } else {
+            await Image.getSize(localFilePath, (width, height) =>
+              this.props.onLoadFinished(width, height)
+            );
+          }
+        }
+      }
     }
 
     async componentWillUnmount() {
+      try {
+        // Track component mount status to avoid calling setState() on unmounted component.
+        this._isMounted = false;
 
-      // Track component mount status to avoid calling setState() on unmounted component.
-      this._isMounted = false;
-
-      // Remove component cache lock on associated image file on component teardown.
-      let fileName = await this.fileSystem.getFileNameFromUrl(traverse(this.props).get(['source', 'uri']));
-      FileSystem.unlockCacheFile(fileName, this.componentId);
-
+        // Remove component cache lock on associated image file on component teardown.
+        let fileName = await this.fileSystem.getFileNameFromUrl(
+          traverse(this.props).get(['source', 'uri'])
+        );
+        FileSystem.unlockCacheFile(fileName, this.componentId);
+      } catch (e) {} // eslint-disable-line no-empty
     }
 
     render() {
-
       // If media loaded, render full image component, else render placeholder.
       if (this.state.localFilePath) {
-
         // Build platform specific file resource uri.
-        const localFileUri = (Platform.OS == 'ios') ? this.state.localFilePath : 'file://' + this.state.localFilePath; // Android requires the traditional 3 prefixed slashes file:/// in a localhost absolute file uri.
-
+        let localFileUri;
+        if (!this.InvalidUrl) {
+          localFileUri =
+            Platform.OS == 'ios' ? this.state.localFilePath : 'file://' + this.state.localFilePath; // Android requires the traditional 3 prefixed slashes file:/// in a localhost absolute file uri.
+        } else {
+          localFileUri = this.state.localFilePath;
+        }
         // Extract props proprietary to this HOC before passing props through.
         let { permanent, ...filteredProps } = this.props; // eslint-disable-line no-unused-vars
-
         let props = Object.assign({}, filteredProps, { source: { uri: localFileUri } });
-        return (<Image {...props} />);
+        return <Image {...props} />;
       } else {
-
         if (this.props.placeholder) {
-          return (<this.props.placeholder.component {...this.props.placeholder.props} />);
+          return <this.props.placeholder.component {...this.props.placeholder.props} />;
         } else if (this.options.defaultPlaceholder) {
-          return (<this.options.defaultPlaceholder.component {...this.options.defaultPlaceholder.props} />);
+          return (
+            <this.options.defaultPlaceholder.component {...this.options.defaultPlaceholder.props} />
+          );
         } else {
-          return (<Image style={this.props.style ? this.props.style : undefined} />);
+          return <Image style={this.props.style ? this.props.style : undefined} />;
         }
-
       }
-
     }
-
   };
-
 }
