@@ -9,12 +9,20 @@
 
 import { Platform } from 'react-native';
 import pathLib from 'path';
-import RNFetchBlob from 'rn-fetch-blob';
+import RNFS from 'react-native-fs';
 import sha1 from 'crypto-js/sha1';
 import URL from 'url-parse';
 
-export class FileSystem {
+// Resolves if 'unlink' resolves or if the file doesn't exist.
+const RNFSUnlinkIfExists = (filename) =>
+  RNFS.exists(filename).then((exists) => {
+    if (exists) {
+      return RNFS.unlink(filename);
+    }
+    return Promise.resolve();
+  });
 
+export class FileSystem {
   /**
    * All FileSystem instances will reference the cacheLock singleton "dictionary" to provide cache file locking in order to prevent concurrency race condition bugs.
    *
@@ -39,7 +47,6 @@ export class FileSystem {
   static cacheLock = {};
 
   static lockCacheFile(fileName, componentId) {
-
     // If file is already locked, add additional component lock, else create initial file lock.
     if (FileSystem.cacheLock[fileName]) {
       FileSystem.cacheLock[fileName][componentId] = true;
@@ -48,21 +55,21 @@ export class FileSystem {
       componentDict[componentId] = true;
       FileSystem.cacheLock[fileName] = componentDict;
     }
-
   }
 
   static unlockCacheFile(fileName, componentId) {
-
     // Delete component lock on cache file
     if (FileSystem.cacheLock[fileName]) {
       delete FileSystem.cacheLock[fileName][componentId];
     }
 
     // If no further component locks remain on cache file, delete filename property from cacheLock dictionary.
-    if (FileSystem.cacheLock[fileName] && Object.keys(FileSystem.cacheLock[fileName]).length === 0) {
+    if (
+      FileSystem.cacheLock[fileName] &&
+      Object.keys(FileSystem.cacheLock[fileName]).length === 0
+    ) {
       delete FileSystem.cacheLock[fileName];
     }
-
   }
 
   constructor(cachePruneTriggerLimit = null, fileDirName = null) {
@@ -87,11 +94,10 @@ export class FileSystem {
    * @private
    */
   _setBaseFilePath(fileDirName = null) {
-    let baseFilePath = (this.os == 'ios') ? RNFetchBlob.fs.dirs.CacheDir : RNFetchBlob.fs.dirs.DocumentDir;
-    baseFilePath +=  '/' + fileDirName + '/';
+    let baseFilePath = this.os == 'ios' ? RNFS.CachesDirectoryPath : RNFS.DocumentDirectoryPath;
+    baseFilePath += '/' + fileDirName + '/';
     return baseFilePath;
   }
-
   /**
    *
    * Pseudo-chroot paths to the baseFilePath set in _setBaseFilePath().
@@ -107,14 +113,13 @@ export class FileSystem {
    * @private
    */
   _validatePath(path, absolute = false) {
-    let resolvedPath = (absolute) ? pathLib.resolve(path) : pathLib.resolve(this.baseFilePath + path); // resolve turns any path into an absolute path (ie: /folder1/folder2/../example.js resolves to /folder1/example.js)
+    let resolvedPath = absolute ? pathLib.resolve(path) : pathLib.resolve(this.baseFilePath + path); // resolve turns any path into an absolute path (ie: /folder1/folder2/../example.js resolves to /folder1/example.js)
 
     if (resolvedPath.substr(0, this.baseFilePath.length) != this.baseFilePath) {
       throw new Error(resolvedPath + ' is not a valid file path.');
     } else {
       return true;
     }
-
   }
 
   /**
@@ -126,7 +131,7 @@ export class FileSystem {
    */
   exists(path) {
     this._validatePath(path);
-    return RNFetchBlob.fs.exists(pathLib.resolve(this.baseFilePath + path));
+    return RNFS.exists(pathLib.resolve(this.baseFilePath + path));
   }
 
   /**
@@ -138,12 +143,11 @@ export class FileSystem {
    * @returns fileName {string} - A SHA1 filename that is unique to the resource located at passed in URL and includes an appropriate extension.
    */
   async getFileNameFromUrl(url) {
-
     const urlParts = new URL(url);
     const urlExt = urlParts.pathname.split('.').pop();
 
     let extension = null;
-    switch(urlExt) {
+    switch (urlExt) {
       case 'png':
       case 'PNG':
         extension = 'png';
@@ -163,7 +167,7 @@ export class FileSystem {
         extension = 'bmp';
         break;
       default:
-        extension = await this.getExtensionFromContentTypeHeader(url);
+        extension = urlExt || 'jpg';
     }
 
     if (!extension) {
@@ -171,59 +175,6 @@ export class FileSystem {
     }
 
     return sha1(url).toString() + '.' + extension;
-
-  }
-
-  /**
-   *
-   * Used to determine appropriate file extension for a remote file that doesn't include
-   * an extension in the url. This method will attempt to determine extension using server
-   * "Content-Type" response header.
-   *
-   * @param url {String} - An absolute url.
-   * @returns extension {string} - A file extension appropriate for remote file.
-   */
-  async getExtensionFromContentTypeHeader(url) {
-
-    let extension = null;
-    let contentType = null;
-
-    // Request "Content-type" header from server.
-    try{
-
-      const response = await fetch(url, {
-        method: 'HEAD'
-      });
-
-      if (response.headers.get('content-type')) {
-        const rawContentType = response.headers.get('content-type'); // headers are case-insensitive, fetch standard is all lower case.
-        contentType = rawContentType.toLowerCase();
-      }
-
-    } catch (error) {
-      console.warn(error); // eslint-disable-line no-console
-    }
-
-    // Use content type header to determine extension.
-    switch(contentType) {
-      case 'image/png':
-        extension = 'png';
-        break;
-      case 'image/gif':
-        extension = 'gif';
-        break;
-      case 'image/jpeg':
-        extension = 'jpg';
-        break;
-      case 'image/bmp':
-        extension = 'bmp';
-        break;
-      default:
-        extension = null;
-    }
-
-    return extension;
-
   }
 
   /**
@@ -236,7 +187,6 @@ export class FileSystem {
    * @returns {Promise} promise that resolves to the local file path of downloaded url file.
    */
   async getLocalFilePathFromUrl(url, permanent) {
-
     let filePath = null;
 
     let fileName = await this.getFileNameFromUrl(url);
@@ -247,22 +197,63 @@ export class FileSystem {
     let exists = await Promise.all([permanentFileExists, cacheFileExists]);
 
     if (exists[0]) {
-
       filePath = this.baseFilePath + 'permanent/' + fileName;
-
     } else if (exists[1]) {
-
       filePath = this.baseFilePath + 'cache/' + fileName;
-
     } else {
-
       let result = await this.fetchFile(url, permanent, null, true); // Clobber must be true to allow concurrent CacheableImage components with same source url (ie: bullet point images).
-      filePath = result.path;
-
+      if (result) {
+        filePath = result.path;
+      }
     }
 
     return filePath;
+  }
 
+  /**
+   *
+   * Convenience method used to get the associated local file path of a web image that has been written to disk.
+   * Copies the given local file to the cache then the local filepath is returned.
+   *
+   * @param local {String} - url of local file file to copy.
+   * @param url {String} - url of file to download.
+   * @param permanent {Boolean} - True persists the file locally indefinitely, false caches the file temporarily (until file is removed during cache pruning).
+   * @param move {Boolean} - True moves the local file, false copies it.
+   * @returns {Promise} promise that resolves to the local file path of downloaded url file.
+   */
+  async getLocalFilePathFromUrlLocal(local, url, permanent, move) {
+    let fileName = await this.getFileNameFromUrl(url);
+
+    let path;
+    try {
+      path = this.baseFilePath + (permanent ? 'permanent' : 'cache') + '/' + fileName;
+      this._validatePath(path, true);
+
+      // Logic here prunes cache directory on "cache" writes to ensure cache doesn't get too large.
+      if (!permanent) {
+        await this.pruneCache();
+      }
+
+      // Hit network and download file to local disk.
+      try {
+        const cacheDirExists = await this.exists(permanent ? 'permanent' : 'cache');
+        if (!cacheDirExists) {
+          await RNFS.mkdir(`${this.baseFilePath}${permanent ? 'permanent' : 'cache'}`);
+        }
+
+        await RNFSUnlinkIfExists(path);
+        const { promise } = move ? RNFS.moveFile(local, path) : RNFS.copyFile(local, path);
+        await promise;
+      } catch (error) {
+        await RNFSUnlinkIfExists(path);
+        // throw error;
+        path = url;
+      }
+    } catch (e) {
+      path = url;
+    }
+
+    return path;
   }
 
   /**
@@ -276,45 +267,50 @@ export class FileSystem {
    * @returns {Promise} promise that resolves to an object that contains the local path of the downloaded file and the filename.
    */
   async fetchFile(url, permanent = false, fileName = null, clobber = false) {
-
-    fileName = fileName || await this.getFileNameFromUrl(url);
-    let path = this.baseFilePath + (permanent ? 'permanent' : 'cache') + '/' + fileName;
-    this._validatePath(path, true);
-
-    // Clobber logic
-    let fileExistsAtPath = await this.exists((permanent ? 'permanent/' : 'cache/') + fileName);
-    if (!clobber && fileExistsAtPath) {
-      throw new Error('A file already exists at ' + path + ' and clobber is set to false.');
-    }
-
-    // Logic here prunes cache directory on "cache" writes to ensure cache doesn't get too large.
-    if (!permanent) {
-      await this.pruneCache();
-    }
-
-    // Hit network and download file to local disk.
-    let result = null;
+    let path;
     try {
+      fileName = fileName || (await this.getFileNameFromUrl(url));
+      path = this.baseFilePath + (permanent ? 'permanent' : 'cache') + '/' + fileName;
+      this._validatePath(path, true);
 
-      result = await RNFetchBlob
-        .config({
-          path: path
-        })
-        .fetch('GET', url);
+      // Clobber logic
+      let fileExistsAtPath = await this.exists((permanent ? 'permanent/' : 'cache/') + fileName);
+      if (!clobber && fileExistsAtPath) {
+        throw new Error('A file already exists at ' + path + ' and clobber is set to false.');
+      }
 
-    } catch(error) {
+      // Logic here prunes cache directory on "cache" writes to ensure cache doesn't get too large.
+      if (!permanent) {
+        await this.pruneCache();
+      }
 
-      // File must be manually removed on download error https://github.com/wkh237/react-native-fetch-blob/issues/331
-      await RNFetchBlob.fs.unlink(path);
-      throw error;
+      // Hit network and download file to local disk.
+      try {
+        const cacheDirExists = await this.exists(permanent ? 'permanent' : 'cache');
+        if (!cacheDirExists) {
+          await RNFS.mkdir(`${this.baseFilePath}${permanent ? 'permanent' : 'cache'}`);
+        }
 
+        const { promise } = RNFS.downloadFile({
+          fromUrl: url,
+          toFile: path,
+        });
+        const response = await promise;
+        if (response.statusCode !== 200) {
+          throw response;
+        }
+      } catch (error) {
+        await RNFSUnlinkIfExists(path);
+        throw error;
+      }
+    } catch (e) {
+      return null;
     }
 
     return {
-      path: result.path(),
-      fileName: pathLib.basename(path)
+      path,
+      fileName: pathLib.basename(path),
     };
-
   }
 
   /**
@@ -325,45 +321,42 @@ export class FileSystem {
    * @returns {Promise}
    */
   async pruneCache() {
-
     // If cache directory does not exist yet there's no need for pruning.
-    if (!await this.exists('cache')) {
+    if (!(await this.exists('cache'))) {
       return;
     }
 
-    // Get blobStatObjects.
-    let lstat = await RNFetchBlob.fs.lstat(this.baseFilePath + 'cache');
+    // Get directory contents
+    let dirContents = await RNFS.readDir(this.baseFilePath + 'cache');
 
-    // Sort blobStatObjects in order of oldest to newest file.
-    lstat.sort((a, b) => {
-      return a.lastModified - b.lastModified;
+    // Sort dirContents in order of oldest to newest file.
+    dirContents.sort((a, b) => {
+      return a.mtime - b.mtime;
     });
 
-    let currentCacheSize = lstat.reduce((cacheSize, blobStatObject) => {
+    let currentCacheSize = dirContents.reduce((cacheSize, blobStatObject) => {
       return cacheSize + parseInt(blobStatObject.size);
     }, 0);
 
     // Prune cache if current cache size is too big.
     if (currentCacheSize > this.cachePruneTriggerLimit) {
-
       let overflowSize = currentCacheSize - this.cachePruneTriggerLimit;
 
       // Keep deleting cached files so long as the current cache size is larger than the size required to trigger cache pruning, or until
       // all cache files have been evaluated.
-      while (overflowSize > 0 && lstat.length) {
-
-        let blobStatObject = lstat.shift();
+      while (overflowSize > 0 && dirContents.length) {
+        let contentFile = dirContents.shift();
 
         // Only prune unlocked files from cache
-        if (!FileSystem.cacheLock[blobStatObject.filename] && this._validatePath('cache/' + blobStatObject.filename)) {
-          overflowSize = overflowSize - parseInt(blobStatObject.size);
-          RNFetchBlob.fs.unlink(this.baseFilePath + 'cache/' + blobStatObject.filename);
+        if (
+          !FileSystem.cacheLock[contentFile.name] &&
+          this._validatePath('cache/' + contentFile.name)
+        ) {
+          overflowSize -= parseInt(contentFile.size);
+          RNFSUnlinkIfExists(this.baseFilePath + 'cache/' + contentFile.name);
         }
-
       }
-
     }
-
   }
 
   /**
@@ -376,14 +369,12 @@ export class FileSystem {
     this._validatePath(path);
 
     try {
-      await RNFetchBlob.fs.unlink(pathLib.resolve(this.baseFilePath + path));
+      await RNFSUnlinkIfExists(pathLib.resolve(this.baseFilePath + path));
       return true;
     } catch (error) {
       return false;
     }
-
   }
-
 }
 
 /**
