@@ -1,67 +1,60 @@
-
 // Define globals for eslint.
 /* global describe it */
-/* global expect require */
+/* global expect */
 
 // Load dependencies
 import should from 'should'; // eslint-disable-line no-unused-vars
 import React from 'react';
 import 'react-native';
 import imageCacheHoc from '../lib/imageCacheHoc';
-import {
-  StyleSheet,
-  View,
-  Text,
-  Image
-} from 'react-native';
+import { StyleSheet, View, Text, Image } from 'react-native';
 import { mockData } from './mockData';
 
 // Note: test renderer must be required after react-native.
 import renderer from 'react-test-renderer';
 
 // Ensure component can mount successfully.
-describe('CacheableImage', function() {
-
+describe('CacheableImage', function () {
   it('renders correctly', () => {
-
     //Mock values for local/remote file request logic.
-    const RNFetchBlob = require('rn-fetch-blob');
-    RNFetchBlob.fs.exists
-      .mockReturnValueOnce(false) // mock not exist in local permanent dir
-      .mockReturnValueOnce(false) // mock not exist in local cache dir
-      .mockReturnValueOnce(false) // mock does not exist to get past clobber
-      .mockReturnValue(true);
+    const RNFS = require('react-native-fs');
+    RNFS.exists
+      .mockResolvedValueOnce(false) // mock not exist in local permanent dir
+      .mockResolvedValueOnce(false) // mock not exist in local cache dir
+      .mockResolvedValueOnce(false) // mock does not exist to get past clobber
+      .mockResolvedValue(true);
 
-    RNFetchBlob.fetch
-      .mockReturnValue({
-        path: () => {
-          return '/this/is/path/to/file.jpg';
-        }
-      });
+    RNFS.downloadFile.mockReturnValue({
+      promise: Promise.resolve({ statusCode: 200 }),
+    });
 
     const CacheableImage = imageCacheHoc(Image);
 
     const styles = StyleSheet.create({
       container: {
+        alignItems: 'center',
         flex: 1,
         justifyContent: 'center',
-        alignItems: 'center'
+      },
+      image: {
+        height: 204,
+        width: 150,
       },
       welcome: {
         fontSize: 20,
+        margin: 10,
         textAlign: 'center',
-        margin: 10
       },
-      image: {
-        width:150,
-        height: 204
-      }
     });
 
     const tree = renderer.create(
       <View style={styles.container}>
         <Text style={styles.welcome}>Test CacheableImage Component</Text>
-        <CacheableImage style={styles.image} source={mockData.mockCacheableImageProps.source} permanent={mockData.mockCacheableImageProps.permanent} />
+        <CacheableImage
+          style={styles.image}
+          source={mockData.mockCacheableImageProps.source}
+          permanent={mockData.mockCacheableImageProps.permanent}
+        />
       </View>
     );
     expect(tree).toMatchSnapshot(); //If UI changes, this snapshot must be updated. See comment below.
@@ -71,61 +64,62 @@ describe('CacheableImage', function() {
      The snapshot should be committed along code changes. When a snapshot test fails, you need to inspect whether it is an intended or unintended change.
      If the change is expected you can invoke Jest with npm test -- -u to overwrite the existing snapshot.
      */
-
   });
 
   it('renders correctly with placeholder prop set', () => {
-
     //Mock values for local/remote file request logic.
-    const RNFetchBlob = require('rn-fetch-blob');
-    RNFetchBlob.fs.exists
-      .mockReturnValueOnce(false) // mock not exist in local permanent dir
-      .mockReturnValueOnce(false) // mock not exist in local cache dir
-      .mockReturnValueOnce(false) // mock does not exist to get past clobber
-      .mockReturnValue(true);
+    const RNFS = require('react-native-fs');
+    RNFS.exists
+      .mockResolvedValueOnce(false) // mock not exist in local permanent dir
+      .mockResolvedValueOnce(false) // mock not exist in local cache dir
+      .mockResolvedValueOnce(false) // mock does not exist to get past clobber
+      .mockResolvedValue(true);
 
-    RNFetchBlob.fetch
-      .mockReturnValue({
-        path: () => {
-          return '/this/is/path/to/file.jpg';
-        }
-      });
+    RNFS.downloadFile.mockReturnValue({
+      promise: Promise.resolve({ statusCode: 200 }),
+    });
 
     const CacheableImage = imageCacheHoc(Image);
 
     const styles = StyleSheet.create({
       container: {
+        alignItems: 'center',
         flex: 1,
         justifyContent: 'center',
-        alignItems: 'center'
+      },
+      image: {
+        height: 204,
+        width: 150,
+      },
+      // eslint-disable-next-line react-native/no-color-literals
+      placeholderImage: {
+        backgroundColor: '#00ffff',
+        height: 204,
+        width: 150,
       },
       welcome: {
         fontSize: 20,
+        margin: 10,
         textAlign: 'center',
-        margin: 10
       },
-      image: {
-        width:150,
-        height: 204
-      },
-      placeholderImage: { // eslint-disable-line react-native/no-color-literals
-        width:150,
-        height: 204,
-        backgroundColor: '#00ffff'
-      }
     });
 
     const propPlaceholder = {
       component: Image,
       props: {
-        style: styles.propPlaceholder
-      }
+        style: styles.propPlaceholder,
+      },
     };
 
     const tree = renderer.create(
       <View style={styles.container}>
         <Text style={styles.welcome}>Test CacheableImage Component</Text>
-        <CacheableImage style={styles.image} source={mockData.mockCacheableImageProps.source} permanent={mockData.mockCacheableImageProps.permanent} placeholder={propPlaceholder} />
+        <CacheableImage
+          style={styles.image}
+          source={mockData.mockCacheableImageProps.source}
+          permanent={mockData.mockCacheableImageProps.permanent}
+          placeholder={propPlaceholder}
+        />
       </View>
     );
     expect(tree).toMatchSnapshot(); //If UI changes, this snapshot must be updated. See comment below.
@@ -135,63 +129,63 @@ describe('CacheableImage', function() {
      The snapshot should be committed along code changes. When a snapshot test fails, you need to inspect whether it is an intended or unintended change.
      If the change is expected you can invoke Jest with npm test -- -u to overwrite the existing snapshot.
      */
-
   });
 
   it('renders correctly with placeholder option set', () => {
-
     //Mock values for local/remote file request logic.
-    const RNFetchBlob = require('rn-fetch-blob');
-    RNFetchBlob.fs.exists
-      .mockReturnValueOnce(false) // mock not exist in local permanent dir
-      .mockReturnValueOnce(false) // mock not exist in local cache dir
-      .mockReturnValueOnce(false) // mock does not exist to get past clobber
-      .mockReturnValue(true);
+    const RNFS = require('react-native-fs');
+    RNFS.exists
+      .mockResolvedValueOnce(false) // mock not exist in local permanent dir
+      .mockResolvedValueOnce(false) // mock not exist in local cache dir
+      .mockResolvedValueOnce(false) // mock does not exist to get past clobber
+      .mockResolvedValue(true);
 
-    RNFetchBlob.fetch
-      .mockReturnValue({
-        path: () => {
-          return '/this/is/path/to/file.jpg';
-        }
-      });
+    RNFS.downloadFile.mockReturnValue({
+      promise: Promise.resolve({ statusCode: 200 }),
+    });
 
     const styles = StyleSheet.create({
       container: {
+        alignItems: 'center',
         flex: 1,
         justifyContent: 'center',
-        alignItems: 'center'
+      },
+      image: {
+        height: 204,
+        width: 150,
+      },
+      // eslint-disable-next-line react-native/no-color-literals
+      optionPlaceholder: {
+        backgroundColor: '#dc143c',
+        height: 204,
+        width: 150,
       },
       welcome: {
         fontSize: 20,
+        margin: 10,
         textAlign: 'center',
-        margin: 10
       },
-      image: {
-        width:150,
-        height: 204
-      },
-      optionPlaceholder: { // eslint-disable-line react-native/no-color-literals
-        width:150,
-        height: 204,
-        backgroundColor: '#dc143c'
-      }
     });
 
     const optionPlaceholder = {
       component: Image,
       props: {
-        style: styles.optionPlaceholder
-      }
+        style: styles.optionPlaceholder,
+      },
     };
 
     const CacheableImage = imageCacheHoc(Image, {
-      defaultPlaceholder: optionPlaceholder
+      defaultPlaceholder: optionPlaceholder,
     });
 
     const tree = renderer.create(
       <View style={styles.container}>
         <Text style={styles.welcome}>Test CacheableImage Component</Text>
-        <CacheableImage style={styles.image} source={mockData.mockCacheableImageProps.source} permanent={mockData.mockCacheableImageProps.permanent} />
+        <CacheableImage
+          style={styles.image}
+          source={mockData.mockCacheableImageProps.source}
+          permanent={mockData.mockCacheableImageProps.permanent}
+        />
       </View>
     );
     expect(tree).toMatchSnapshot(); //If UI changes, this snapshot must be updated. See comment below.
@@ -201,7 +195,5 @@ describe('CacheableImage', function() {
      The snapshot should be committed along code changes. When a snapshot test fails, you need to inspect whether it is an intended or unintended change.
      If the change is expected you can invoke Jest with npm test -- -u to overwrite the existing snapshot.
      */
-
   });
-
 });
